ai-service/requirements.txt
openai
langchain
langchain-openai
streamlit
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv